renaming binpacker to shippo-packer
